requirements.txt
pydantic~=1.10.7
bson~=0.5.10
pymongo~=4.3.3
fastapi~=0.95.1
passlib~=1.7.4
motor~=3.1.2
python-dotenv==1.0.0
python-jose~=3.3.0
uvicorn==0.21.1